slf4j v2 bridge - wip
