+++ slf4j2-bridge/src/main/java/module-info.java
@@ -1,4 +1,4 @@
 module org.slf4j.zio {
     requires org.slf4j;
-    provides org.slf4j.spi.SLF4JServiceProvider with org.slf4j.zio.ZioSLF4JServiceProvider;
+    provides org.slf4j.spi.SLF4JServiceProvider with zio.logging.slf4j.bridge.ZioSLF4JServiceProvider;
 }
