+++ slf4j2-bridge/src/main/java/zio/logging/slf4j/bridge/LoggerRuntime.java
@@ -1,8 +1,8 @@
-package org.slf4j.zio;
+package zio.logging.slf4j.bridge;
 
 import org.slf4j.Marker;
 import org.slf4j.event.Level;
 
-public interface LoggerRuntime {
+interface LoggerRuntime {
     void log(String name, Level level, Marker marker, String messagePattern, Object[] arguments, Throwable throwable);
 }
