+++ slf4j2-bridge/src/main/java/zio/logging/slf4j/bridge/ZioSLF4JServiceProvider.java
@@ -1,4 +1,4 @@
-package org.slf4j.zio;
+package zio.logging.slf4j.bridge;
 
 
 import org.slf4j.ILoggerFactory;
