+++ slf4j2-bridge/src/main/resources/META-INF/services/org.slf4j.spi.SLF4JServiceProvider
@@ -1,1 +1,1 @@
-org.slf4j.zio.ZioSLF4JServiceProvider
+zio.logging.slf4j.bridge.ZioSLF4JServiceProvider
